Adding ability to preview several all days.
Using buttons to toggle the days.

The current day is marked with the class "active" (and has a blue
colored number).
The current preview day has inset borders and is disabled.

If a day is switched, the page is reloaded (window.location.href).
If the day matches the current day, the GET parameters are removed from
the location. Otherwise the ?day=X parameter is set.