drop password from the upload
Instead, enforce a quota on the files/ folder. (outside this git)
The files are reviewed before release and if someone spams the upload
we can still reintroduce the password.

Also the previous password lead to confusion ("lecture" or "lecturer"?),
so I changed it as well in the commented out section